# RealEstateAgency.WebApplication/Api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Client API",
    "description": "API для работы с клиентами",
    "version": "v1"
  },
  "paths": {
...
    "/api/RealEstates": {
      "get": {
        "tags": [
          "RealEstates"
        ],
        "summary": "получить список всех объектов недвижимости",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RealEstateGetDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RealEstateGetDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RealEstateGetDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "RealEstates"
        ],
        "summary": "Добавить новый объект недвижимости.",
        "requestBody": {
          "description": "Значения для свойства `PropertyType` могут быть следующими:\r\n<list type=\"bullet\"><item><term> Residential</term></item><item><term> Uninhabitable</term></item></list>",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RealEstateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RealEstateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RealEstateDto"
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/RealEstates/{id}": {
      "get": {
        "tags": [
          "RealEstates"
        ],
        "summary": "получить объект недвижимости по его идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "идентификатор объекта недвижимости",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RealEstateGetDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RealEstateGetDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RealEstateGetDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "RealEstates"
        ],
        "summary": "обновить данные существующего объекта недвижимости",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "идентификатор объекта недвижимости",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "объект недвижимости с обновленными данными",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RealEstateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RealEstateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RealEstateDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "RealEstates"
        ],
        "summary": "удалить объект недвижимости по его идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "идентификатор объекта недвижимости",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {}
      },
      "RealEstateDto": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "nullable": true
          },
          "square": {
            "type": "number",
            "format": "double"
          },
          "numberOfRooms": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RealEstateGetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "square": {
            "type": "number",
            "format": "double"
          },
          "numberOfRooms": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}